+++ CHANGES.md
@@ -19,7 +19,7 @@ To be released.
 
  -  Added `TurnClient.BindProxies()` method. [[#756], [#868]]
  -  Added `ActionEvaluation.Exception` property.  [[#860], [[#875]]]
- -  Added `allowRender` optional parameter to `BlockChain<T>()` constructor.  [[#883]]
+ -  Added `render` optional parameter to `BlockChain<T>()` constructor.  [[#883]]
 
 ### Behavioral changes